+++ database/migrations/2024_10_11_154703_create_temp_machine_statuses_table.php
@@ -30,7 +30,7 @@ class CreateTempMachineStatusesTable extends Migration
             $table->dateTime('shift_start_datetime')->nullable();
             $table->dateTime('shift_end_datetime')->nullable();
             $table->date('machine_date')->nullable();
-            $table->integer('status')->comment('1->Start, 0->Stop');
+            $table->unsignedBigInteger('status')->comment('1->Start, 0->Stop');
             $table->unsignedBigInteger('machine_status_id');
             $table->longText('machine_log');
             $table->integer('created_by')->nullable();
